added working file upload with orm

+++ universitysearch/src/main/java/universitysearch/FileManager.java
@@ -13,14 +13,11 @@ public class FileManager extends DBManager {
 	}
 	
 	/* Method to add a file to the database */
-	public Integer addFile(String fName, String fPath, String fDesc, String fHash, int fSize, int fOwn){
+	public Integer addFile(String fName, String fPath, String fDesc, String fHash, long fSize, int fOwn){
 		Session session = factory.openSession();
 	    Transaction tx = null;
 		Integer userID = null;
-		//User user = new User(email, pass, fName, lName, isPr, isAd, emailVer);
-		//String addUser = "INSERT INTO `universitysearch`.`users` (`email`, `password`, `first_name`, `last_name`) VALUES ('" + 
-		//email + "', '"+ pass + "', '" + fName + "', '" + lName + ")" ; 
-        
+
 		try{
 	         tx = session.beginTransaction();
 	         File file = new File (fName, fPath, fDesc, fHash, fSize, fOwn);
